refactor(vimac64b): reconstruct vimac64b module
	refactor(vimac64b): reconstruct vimac64b module
	feat(golden model): support vimac
	feat(vecsimtop):    add vimac interfaces
